+++ packages/asc-ui/src/components/Card/CardContainerStyle.ts
@@ -3,13 +3,9 @@ import CardStyle from './CardStyle'
 
 const CardContainerStyle = styled.div`
   width: 100%;
-  padding: 24px;
 
   ${CardStyle} {
     margin-bottom: 12px;
-  }
-  ${/* sc-selector */ CardStyle}:last-of-type {
-    margin-bottom: 0;
   }
 `