+++ packages/asc-ui/src/components/Paragraph/Paragraph.stories.tsx
@@ -1,6 +1,7 @@
 import * as React from 'react'
 import { storiesOf } from '@storybook/react'
 import Paragraph from './Paragraph'
+import CompactThemeProvider from '../../theme/CompactThemeProvider'
 
 storiesOf('Atoms/Typography/Paragraph', module)
   .add('default style', () => (
@@ -29,4 +30,15 @@ storiesOf('Atoms/Typography/Paragraph', module)
         suscipit.
       </Paragraph>
     </>
+  ))
+  .add('compact paragraph', () => (
+    <CompactThemeProvider>
+      <div style={{ width: '300px' }}>
+        <Paragraph>
+          Lorem ipsum dolor sit amet, consectetur adipisicing elit. Debitis
+          deleniti dicta doloremque eveniet explicabo id ipsum omnis placeat
+          rerum suscipit.
+        </Paragraph>
+      </div>
+    </CompactThemeProvider>
   ))